# openapi/choruz.yaml
openapi: 3.1.0
info:
  title: Choruz API
  version: 0.1.0
  description: >
    Implementation route inventory for the local Choruz API Gateway. Routes
    classified under x-choruz-route-classes.humanOperatorInternal or
    x-choruz-route-classes.operationalInternal are documented so the console and
    operator tooling contract is visible; they are not ordinary third-party API
    integration endpoints.
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/companies/{companyId}/harness-accounts/{accountId}/probe:
    parameters:
      - name: companyId
        in: path
        required: true
        schema:
          type: string
      - name: accountId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: probeHarnessAccount
      summary: Read the account's identity, models and exact quota on the device that holds its login and store the snapshot
      responses:
        "204":
          description: Snapshot stored
        "409":
          description: The probe failed; the reason is recorded on the account
...
  /v1/conversations/{conversationId}/messages/{messageId}:
    parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: string
      - name: messageId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getMessage
      responses:
        "200":
          description: Message
...
  /v1/runtime-hosts/{hostId}/operations:
    parameters:
      - name: hostId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: executeRuntimeHostOperation
      description: Runs a device-owned operation. drivers.inspect without driver_type reports installed harnesses; with driver_type it reads that device's default-profile model catalog.
      responses:
        "200":
          description: Result returned by the selected runtime host
...
components:
  securitySchemes:
    RuntimeHostToken:
      type: apiKey
      in: header
      name: x-choruz-host-token
      description: The token a runtime host received when its pairing code was redeemed
...